+++ js/dataset/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/dataset
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @gradio/atoms@0.3.1
+  - @gradio/upload@0.5.2
+
 ## 0.1.8
 
 ### Patch Changes