+++ js/tootils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/tootils
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @gradio/statustracker@0.4.1
+
 ## 0.1.4
 
 ### Patch Changes